Пересморел подход к разбивке. Не работало отображение фотографий. Упростил модули.
