--- src/constants.js
@@ -1,3 +1,0 @@
-const POINT_COUNT = 3;
-
-export { POINT_COUNT };
