--- src/view/page-main/list-container-item-view.js
@@ -1,21 +1,0 @@
-import { createElement } from '../../render.js';
-
-const createTripListContainerItemTemplate = () => '<li class="trip-events__item"></li>';
-
-export default class ListContainerItemView {
-  getTemplate() {
-    return createTripListContainerItemTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
